Wait for the lock release during ip move
In the case of ip move the source node may be still releasing
the state that we need to acquire. So let us be little patient
to STATE_LOCK_CONFLICT errors if the lock is a reclaim and we are
in grace and retry it.

Port from 1.5: f63fc49e7d4c76fcaccf8abd1984e77fe16197cb

Signed-off-by: Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com>